Added basic security for api endpoints (need to update) and added favicon

+++ src/routes/api/v1/delete/booking/+server.js
@@ -3,6 +3,13 @@ import { deleteReservation } from '$lib/pocketbase';
 import { json } from '@sveltejs/kit';
 
 export async function POST({ request, cookies }) {
+	const origin = request.headers.get('Origin');
+	const allowedOrigins = [import.meta.env.VITE_42API_URL];
+
+	if (!origin || !allowedOrigins.includes(origin)) {
+			throw error(403, 'Origin not allowed');
+	}
+
 	const { id } = await request.json();
 	if (id) {
 		deleteReservation(id);
